digital clock using html ,css and javascript.
HTML: In this section, we have a dummy time in the format of “HH:MM:SS” wrapped inside a “div” tag. 

CSS: For CSS, we have just aligned our clock to the center of the page. Other than that, it is just some font-size and width which you can adjust according to your need. 

JavaScript: For JavaScript, follow the below-given steps.

Step 1: Create a function “showTime”.
Step 2: Create an instance of the Date object.
Step 3: Using the methods of the Date object get “hours”, “minute” and “seconds”.
Step 4: Set AM/PM depending on the hour value. The Date object works on the 24-hour format so we change the hour back to 1 when it gets larger than 12. The AM/PM also changes according to that.
Step 5: Now make a string using the same HH:MM:SS format changing the hour, minute, and second value with the values, we get from Date object methods.
Step 6: Now replace the string variable in the “div” using the innerHTML property.
Step 7: To call the function every second use the setInterval() method and set the time-interval as 1000ms which is equal to 1s.
Step 8: Now call the function at the end to start the function at the exact reloading/rendering time as setInterval() will call first after 1s of rendering